fixed merge rows
# C# did not add floats correctly. Switched to decimal
# fixed too many addColumn for "merge rows".
# Less  columns on merge rows.  Only create a column if the target value is not empty.

+++ DataTableConverter/Classes/MergeRowsInfo.cs
@@ -9,13 +9,13 @@ namespace DataTableConverter.Classes
 {
     class MergeRowsInfo
     {
-        internal Dictionary<string,float> CountDict;
+        internal Dictionary<string, decimal> CountDict;
         internal int RowsMerged = 0;
         internal DataRow Row;
 
         public MergeRowsInfo(DataRow row)
         {
-            CountDict = new Dictionary<string, float>();
+            CountDict = new Dictionary<string, decimal>();
             Row = row;
         }
     }
